fixes to naked call tests
