+++ test/integration-tests/nakedCallPreExpiry.test.ts
@@ -1,7 +1,7 @@
 import {
   MockERC20Instance,
   MarginCalculatorInstance,
-  MockAddressBookInstance,
+  AddressBookInstance,
   MockOracleInstance,
   OtokenInstance,
   ControllerInstance,
@@ -15,7 +15,7 @@ import BigNumber from 'bignumber.js'
 import Reverter from '../Reverter'
 
 const {expectRevert, time} = require('@openzeppelin/test-helpers')
-const AddressBook = artifacts.require('MockAddressBook.sol')
+const AddressBook = artifacts.require('AddressBook.sol')
 const MockOracle = artifacts.require('MockOracle.sol')
 const Otoken = artifacts.require('Otoken.sol')
 const MockERC20 = artifacts.require('MockERC20.sol')
@@ -38,14 +38,15 @@ enum ActionType {
   Call,
 }
 
-contract('Naked Put Option flow', ([admin, accountOwner1, accountOperator1, buyer]) => {
+contract('Naked Call Option flow', ([admin, accountOwner1, accountOperator1, buyer]) => {
   const reverter = new Reverter(web3)
 
   let expiry: number
 
-  let addressBook: MockAddressBookInstance
+  let addressBook: AddressBookInstance
   let calculator: MarginCalculatorInstance
-  let controller: ControllerInstance
+  let controllerProxy: ControllerInstance
+  let controllerImplementation: ControllerInstance
   let marginPool: MarginPoolInstance
 
   // whitelist module mock
@@ -71,12 +72,12 @@ contract('Naked Put Option flow', ([admin, accountOwner1, accountOperator1, buye
     // initiate addressbook first.
     addressBook = await AddressBook.new()
     // setup calculator
-    calculator = await MarginCalculator.new()
-    await calculator.init(addressBook.address)
+    calculator = await MarginCalculator.new(addressBook.address)
     // setup margin pool
     marginPool = await MarginPool.new(addressBook.address)
-    // setup controller module
-    controller = await Controller.new(addressBook.address)
+    // setup controllerProxy module
+    controllerImplementation = await Controller.new(addressBook.address)
+    controllerProxy = await Controller.new(addressBook.address)
     // setup mock Oracle module
     oracle = await MockOracle.new(addressBook.address)
     // setup mock whitelist module
@@ -88,12 +89,15 @@ contract('Naked Put Option flow', ([admin, accountOwner1, accountOperator1, buye
     dai = await MockERC20.new('DAI', 'DAI', 18)
     weth = await MockERC20.new('WETH', 'WETH', 18)
 
-    // TODO: setup address book
+    // setup address book
     await addressBook.setOracle(oracle.address)
-    await addressBook.setController(controller.address)
+    await addressBook.setController(controllerImplementation.address)
     await addressBook.setMarginCalculator(calculator.address)
     await addressBook.setWhitelist(whitelist.address)
     await addressBook.setMarginPool(marginPool.address)
+
+    const controllerProxyAddress = await addressBook.getController()
+    controllerProxy = await Controller.at(controllerProxyAddress)
 
     ethCall = await Otoken.new()
     await ethCall.init(
@@ -116,11 +120,11 @@ contract('Naked Put Option flow', ([admin, accountOwner1, accountOperator1, buye
     // have the user approve all the weth transfers
     weth.approve(marginPool.address, '10000000000000000000000', {from: accountOwner1})
 
-    const vaultCounterBefore = new BigNumber(await controller.getAccountVaultCounter(accountOwner1))
+    const vaultCounterBefore = new BigNumber(await controllerProxy.getAccountVaultCounter(accountOwner1))
     vaultCounter = vaultCounterBefore.toNumber() + 1
   })
 
-  describe('Integration test: Sell a naked short call and close it before expiry', () => {
+  describe('Integration test: Sell a naked call and close it before expiry', () => {
     it('Seller should be able to open a short call option', async () => {
       // Keep track of balances before
       const ownerWethBalanceBefore = new BigNumber(await weth.balanceOf(accountOwner1))
@@ -129,7 +133,7 @@ contract('Naked Put Option flow', ([admin, accountOwner1, accountOperator1, buye
       const marginPoolOtokenSupplyBefore = new BigNumber(await ethCall.totalSupply())
 
       // Check that we start at a valid state
-      const vaultBefore = await controller.getVault(accountOwner1, vaultCounter)
+      const vaultBefore = await controllerProxy.getVault(accountOwner1, vaultCounter)
       const vaultStateBefore = await calculator.getExcessCollateral(vaultBefore)
       assert.equal(vaultStateBefore[0].toString(), '0')
       assert.equal(vaultStateBefore[1], true)
@@ -180,7 +184,7 @@ contract('Naked Put Option flow', ([admin, accountOwner1, accountOperator1, buye
         },
       ]
 
-      await controller.operate(actionArgs, {from: accountOwner1})
+      await controllerProxy.operate(actionArgs, {from: accountOwner1})
 
       // keep track of balances after
       const ownerWethBalanceAfter = new BigNumber(await weth.balanceOf(accountOwner1))
@@ -212,7 +216,7 @@ contract('Naked Put Option flow', ([admin, accountOwner1, accountOperator1, buye
       )
 
       // Check that we end at a valid state
-      const vaultAfter = await controller.getVault(accountOwner1, vaultCounter)
+      const vaultAfter = await controllerProxy.getVault(accountOwner1, vaultCounter)
       const vaultStateAfter = await calculator.getExcessCollateral(vaultAfter)
       assert.equal(vaultStateAfter[0].toString(), '0')
       assert.equal(vaultStateAfter[1], true)
@@ -265,7 +269,7 @@ contract('Naked Put Option flow', ([admin, accountOwner1, accountOperator1, buye
         },
       ]
 
-      await controller.operate(actionArgs, {from: accountOwner1})
+      await controllerProxy.operate(actionArgs, {from: accountOwner1})
 
       // keep track of balances after
       const ownerWethBalanceAfter = new BigNumber(await weth.balanceOf(accountOwner1))
@@ -286,7 +290,7 @@ contract('Naked Put Option flow', ([admin, accountOwner1, accountOperator1, buye
       )
 
       // Check that there is excess margin
-      const vaultAfter = await controller.getVault(accountOwner1, vaultCounter)
+      const vaultAfter = await controllerProxy.getVault(accountOwner1, vaultCounter)
       const vaultStateAfter = await calculator.getExcessCollateral(vaultAfter)
       assert.equal(
         vaultStateAfter[0].toString(),
@@ -339,7 +343,7 @@ contract('Naked Put Option flow', ([admin, accountOwner1, accountOperator1, buye
         },
       ]
 
-      await controller.operate(actionArgs, {from: accountOwner1})
+      await controllerProxy.operate(actionArgs, {from: accountOwner1})
 
       // keep track of balances after
       const ownerWethBalanceAfter = new BigNumber(await weth.balanceOf(accountOwner1))
@@ -360,7 +364,7 @@ contract('Naked Put Option flow', ([admin, accountOwner1, accountOperator1, buye
       )
 
       // Check that we end at a valid state with no extra collateral
-      const vaultAfter = await controller.getVault(accountOwner1, vaultCounter)
+      const vaultAfter = await controllerProxy.getVault(accountOwner1, vaultCounter)
       const vaultStateAfter = await calculator.getExcessCollateral(vaultAfter)
       assert.equal(vaultStateAfter[0].toString(), '0')
       assert.equal(vaultStateAfter[1], true)
@@ -407,7 +411,7 @@ contract('Naked Put Option flow', ([admin, accountOwner1, accountOperator1, buye
         },
       ]
       // TODO: Revert message to be updated
-      await expectRevert.unspecified(controller.operate(actionArgs, {from: accountOwner1}))
+      await expectRevert.unspecified(controllerProxy.operate(actionArgs, {from: accountOwner1}))
     })
 
     it('should be able to transfer long otokens to another address', async () => {
@@ -434,7 +438,7 @@ contract('Naked Put Option flow', ([admin, accountOwner1, accountOperator1, buye
       ethCall.transfer(accountOwner1, createScaledUint256(optionsAmount, 18), {from: buyer})
     })
 
-    xit('should be able to close out the short position', async () => {
+    it('should be able to close out the short position', async () => {
       // Keep track of balances before
       const ownerWethBalanceBefore = new BigNumber(await weth.balanceOf(accountOwner1))
       const marginPoolWethBalanceBefore = new BigNumber(await weth.balanceOf(marginPool.address))
@@ -442,7 +446,7 @@ contract('Naked Put Option flow', ([admin, accountOwner1, accountOperator1, buye
       const marginPoolOtokenSupplyBefore = new BigNumber(await ethCall.totalSupply())
 
       // Check that we start at a valid state
-      const vaultBefore = await controller.getVault(accountOwner1, vaultCounter)
+      const vaultBefore = await controllerProxy.getVault(accountOwner1, vaultCounter)
       const vaultStateBefore = await calculator.getExcessCollateral(vaultBefore)
       assert.equal(vaultStateBefore[0].toString(), '0')
       assert.equal(vaultStateBefore[1], true)
@@ -470,7 +474,7 @@ contract('Naked Put Option flow', ([admin, accountOwner1, accountOperator1, buye
         },
       ]
 
-      await controller.operate(actionArgs, {from: accountOwner1})
+      await controllerProxy.operate(actionArgs, {from: accountOwner1})
 
       // keep track of balances after
       const ownerWethBalanceAfter = new BigNumber(await weth.balanceOf(accountOwner1))
@@ -502,7 +506,7 @@ contract('Naked Put Option flow', ([admin, accountOwner1, accountOperator1, buye
       )
 
       // Check that we end at a valid state
-      const vaultAfter = await controller.getVault(accountOwner1, vaultCounter)
+      const vaultAfter = await controllerProxy.getVault(accountOwner1, vaultCounter)
       const vaultStateAfter = await calculator.getExcessCollateral(vaultAfter)
       assert.equal(vaultStateAfter[0].toString(), '0')
       assert.equal(vaultStateAfter[1], true)
@@ -529,303 +533,6 @@ contract('Naked Put Option flow', ([admin, accountOwner1, accountOperator1, buye
         '0',
         'Incorrect amount of collateral stored in the vault',
       )
-    })
-  })
-
-  describe('Integration test: Sell a naked short call and close it after expiry', () => {
-    describe('OTM Tests', async () => {
-      before('revert to state where options have been created', async () => {
-        await reverter.revert()
-      })
-
-      xit('Seller: close an OTM position after expiry', async () => {
-        // Set the oracle price
-        if ((await time.latest()) < expiry) {
-          await time.increaseTo(expiry + 2)
-        }
-        await oracle.setIsFinalized(weth.address, expiry, true)
-        const strikePriceChange = 100
-        const expirySpotPrice = strikePrice - strikePriceChange
-        await oracle.setExpiryPrice(weth.address, expiry, createScaledUint256(expirySpotPrice, 18))
-
-        // Keep track of balances before
-        const ownerWethBalanceBefore = new BigNumber(await weth.balanceOf(accountOwner1))
-        const marginPoolWethBalanceBefore = new BigNumber(await weth.balanceOf(marginPool.address))
-        const ownerOtokenBalanceBefore = new BigNumber(await ethCall.balanceOf(accountOwner1))
-        const marginPoolOtokenSupplyBefore = new BigNumber(await ethCall.totalSupply())
-
-        // Check that we start at a valid state
-        const vaultBefore = await controller.getVault(accountOwner1, vaultCounter)
-        const vaultStateBefore = await calculator.getExcessCollateral(vaultBefore)
-        assert.equal(vaultStateBefore[0].toString(), '0')
-        assert.equal(vaultStateBefore[1], true)
-
-        const actionArgs = [
-          {
-            actionType: ActionType.SettleVault,
-            owner: accountOwner1,
-            sender: accountOwner1,
-            asset: ZERO_ADDR,
-            vaultId: vaultCounter,
-            amount: '0',
-            index: '0',
-            data: ZERO_ADDR,
-          },
-        ]
-
-        await controller.operate(actionArgs, {from: accountOwner1})
-
-        // keep track of balances after
-        const ownerWethBalanceAfter = new BigNumber(await weth.balanceOf(accountOwner1))
-        const marginPoolWethBalanceAfter = new BigNumber(await weth.balanceOf(marginPool.address))
-
-        const ownerOtokenBalanceAfter = new BigNumber(await ethCall.balanceOf(accountOwner1))
-        const marginPoolOtokenSupplyAfter = new BigNumber(await ethCall.totalSupply())
-
-        // check balances before and after changed as expected
-        assert.equal(
-          ownerWethBalanceBefore
-            .plus(createScaledUint256(collateralAmount, (await weth.decimals()).toNumber()))
-            .toString(),
-          ownerWethBalanceAfter.toString(),
-        )
-        assert.equal(
-          marginPoolWethBalanceBefore
-            .minus(createScaledUint256(collateralAmount, (await weth.decimals()).toNumber()))
-            .toString(),
-          marginPoolWethBalanceAfter.toString(),
-        )
-        assert.equal(ownerOtokenBalanceBefore.toString(), ownerOtokenBalanceAfter.toString())
-        assert.equal(marginPoolOtokenSupplyBefore.toString(), marginPoolOtokenSupplyAfter.toString())
-
-        // Check that we end at a valid state
-        const vaultAfter = await controller.getVault(accountOwner1, vaultCounter)
-        const vaultStateAfter = await calculator.getExcessCollateral(vaultAfter)
-        assert.equal(vaultStateAfter[0].toString(), '0')
-        assert.equal(vaultStateAfter[1], true)
-
-        // Check the vault balances stored in the contract
-        assert.equal(vaultAfter.shortOtokens.length, 1, 'Length of the short otoken array in the vault is incorrect')
-        assert.equal(vaultAfter.collateralAssets.length, 1, 'Length of the collateral array in the vault is incorrect')
-        assert.equal(vaultAfter.longOtokens.length, 0, 'Length of the long otoken array in the vault is incorrect')
-
-        assert.equal(vaultAfter.shortOtokens[0], ZERO_ADDR, 'Incorrect short otoken in the vault')
-        assert.equal(vaultAfter.collateralAssets[0], ZERO_ADDR, 'Incorrect collateral asset in the vault')
-
-        assert.equal(vaultAfter.shortAmounts.length, 1, 'Length of the short amounts array in the vault is incorrect')
-        assert.equal(
-          vaultAfter.collateralAmounts.length,
-          1,
-          'Length of the collateral amounts array in the vault is incorrect',
-        )
-        assert.equal(vaultAfter.longAmounts.length, 0, 'Length of the long amounts array in the vault is incorrect')
-
-        assert.equal(vaultAfter.shortAmounts[0].toString(), '0', 'Incorrect amount of short stored in the vault')
-        assert.equal(
-          vaultAfter.collateralAmounts[0].toString(),
-          '0',
-          'Incorrect amount of collateral stored in the vault',
-        )
-      })
-
-      xit('Buyer: exercise OTM call option after expiry', async () => {
-        // owner sells their call option
-        ethCall.transfer(buyer, createScaledUint256(optionsAmount, 18), {from: accountOwner1})
-
-        // Keep track of balances before
-        const ownerWethBalanceBefore = new BigNumber(await weth.balanceOf(buyer))
-        const marginPoolWethBalanceBefore = new BigNumber(await weth.balanceOf(marginPool.address))
-        const ownerOtokenBalanceBefore = new BigNumber(await ethCall.balanceOf(buyer))
-        const marginPoolOtokenSupplyBefore = new BigNumber(await ethCall.totalSupply())
-
-        const actionArgs = [
-          {
-            actionType: ActionType.Exercise,
-            owner: buyer,
-            sender: buyer,
-            asset: ethCall.address,
-            vaultId: '0',
-            amount: createScaledUint256(optionsAmount, 18),
-            index: '0',
-            data: ZERO_ADDR,
-          },
-        ]
-
-        await ethCall.approve(marginPool.address, createScaledUint256(optionsAmount, 18), {from: buyer})
-        await controller.operate(actionArgs, {from: buyer})
-
-        // keep track of balances after
-        const ownerWethBalanceAfter = new BigNumber(await weth.balanceOf(buyer))
-        const marginPoolWethBalanceAfter = new BigNumber(await weth.balanceOf(marginPool.address))
-        const ownerOtokenBalanceAfter = new BigNumber(await ethCall.balanceOf(buyer))
-        const marginPoolOtokenSupplyAfter = new BigNumber(await ethCall.totalSupply())
-
-        // check balances before and after changed as expected
-        assert.equal(ownerWethBalanceBefore.toString(), ownerWethBalanceAfter.toString())
-        assert.equal(marginPoolWethBalanceBefore.toString(), marginPoolWethBalanceAfter.toString())
-        assert.equal(
-          ownerOtokenBalanceBefore.minus(createScaledUint256(optionsAmount, 18)).toString(),
-          ownerOtokenBalanceAfter.toString(),
-        )
-        assert.equal(
-          marginPoolOtokenSupplyBefore.minus(createScaledUint256(optionsAmount, 18)).toString(),
-          marginPoolOtokenSupplyAfter.toString(),
-        )
-      })
-    })
-
-    describe('ITM Tests', async () => {
-      before('revert to state where options have been created', async () => {
-        await reverter.revert()
-      })
-
-      xit('Seller: close an ITM position after expiry', async () => {
-        // Set the oracle price
-        if ((await time.latest()) < expiry) {
-          await time.increaseTo(expiry + 2)
-        }
-        await oracle.setIsFinalized(weth.address, expiry, true)
-        const strikePriceChange = 100
-        const expirySpotPrice = strikePrice + strikePriceChange
-        await oracle.setExpiryPrice(weth.address, expiry, createScaledUint256(expirySpotPrice, 18))
-
-        // Keep track of balances before
-        const ownerWethBalanceBefore = new BigNumber(await weth.balanceOf(accountOwner1))
-        const marginPoolWethBalanceBefore = new BigNumber(await weth.balanceOf(marginPool.address))
-        const ownerOtokenBalanceBefore = new BigNumber(await ethCall.balanceOf(accountOwner1))
-        const marginPoolOtokenSupplyBefore = new BigNumber(await ethCall.totalSupply())
-
-        // Check that we start at a valid state
-        const vaultBefore = await controller.getVault(accountOwner1, vaultCounter)
-        const vaultStateBefore = await calculator.getExcessCollateral(vaultBefore)
-        assert.equal(vaultStateBefore[0].toString(), '0')
-        assert.equal(vaultStateBefore[1], true)
-
-        const actionArgs = [
-          {
-            actionType: ActionType.SettleVault,
-            owner: accountOwner1,
-            sender: accountOwner1,
-            asset: ZERO_ADDR,
-            vaultId: vaultCounter,
-            amount: '0',
-            index: '0',
-            data: ZERO_ADDR,
-          },
-        ]
-
-        await controller.operate(actionArgs, {from: accountOwner1})
-
-        // keep track of balances after
-        const ownerWethBalanceAfter = new BigNumber(await weth.balanceOf(accountOwner1))
-        const marginPoolWethBalanceAfter = new BigNumber(await weth.balanceOf(marginPool.address))
-
-        const ownerOtokenBalanceAfter = new BigNumber(await ethCall.balanceOf(accountOwner1))
-        const marginPoolOtokenSupplyAfter = new BigNumber(await ethCall.totalSupply())
-
-        const collateralPayout = collateralAmount - (strikePriceChange * optionsAmount) / expirySpotPrice
-
-        // check balances before and after changed as expected
-        assert.equal(
-          ownerWethBalanceBefore
-            .plus(createScaledUint256(collateralPayout, (await weth.decimals()).toNumber()))
-            .toString(),
-          ownerWethBalanceAfter.toString(),
-        )
-        assert.equal(
-          marginPoolWethBalanceBefore
-            .minus(createScaledUint256(collateralPayout, (await weth.decimals()).toNumber()))
-            .toString(),
-          marginPoolWethBalanceAfter.toString(),
-        )
-        assert.equal(ownerOtokenBalanceBefore.toString(), ownerOtokenBalanceAfter.toString())
-        assert.equal(marginPoolOtokenSupplyBefore.toString(), marginPoolOtokenSupplyAfter.toString())
-
-        // Check that we end at a valid state
-        const vaultAfter = await controller.getVault(accountOwner1, vaultCounter)
-        const vaultStateAfter = await calculator.getExcessCollateral(vaultAfter)
-        assert.equal(vaultStateAfter[0].toString(), '0')
-        assert.equal(vaultStateAfter[1], true)
-
-        // Check the vault balances stored in the contract
-        assert.equal(vaultAfter.shortOtokens.length, 1, 'Length of the short otoken array in the vault is incorrect')
-        assert.equal(vaultAfter.collateralAssets.length, 1, 'Length of the collateral array in the vault is incorrect')
-        assert.equal(vaultAfter.longOtokens.length, 0, 'Length of the long otoken array in the vault is incorrect')
-
-        assert.equal(vaultAfter.shortOtokens[0], ZERO_ADDR, 'Incorrect short otoken in the vault')
-        assert.equal(vaultAfter.collateralAssets[0], ZERO_ADDR, 'Incorrect collateral asset in the vault')
-
-        assert.equal(vaultAfter.shortAmounts.length, 1, 'Length of the short amounts array in the vault is incorrect')
-        assert.equal(
-          vaultAfter.collateralAmounts.length,
-          1,
-          'Length of the collateral amounts array in the vault is incorrect',
-        )
-        assert.equal(vaultAfter.longAmounts.length, 0, 'Length of the long amounts array in the vault is incorrect')
-
-        assert.equal(vaultAfter.shortAmounts[0].toString(), '0', 'Incorrect amount of short stored in the vault')
-        assert.equal(
-          vaultAfter.collateralAmounts[0].toString(),
-          '0',
-          'Incorrect amount of collateral stored in the vault',
-        )
-      })
-
-      xit('Buyer: exercise ITM call option after expiry', async () => {
-        // owner sells their call option
-        ethCall.transfer(buyer, createScaledUint256(optionsAmount, 18), {from: accountOwner1})
-        // oracle orice decreases
-        const strikePriceChange = 100
-        const expirySpotPrice = strikePrice + strikePriceChange
-
-        // Keep track of balances before
-        const ownerWethBalanceBefore = new BigNumber(await weth.balanceOf(buyer))
-        const marginPoolWethBalanceBefore = new BigNumber(await weth.balanceOf(marginPool.address))
-        const ownerOtokenBalanceBefore = new BigNumber(await ethCall.balanceOf(buyer))
-        const marginPoolOtokenSupplyBefore = new BigNumber(await ethCall.totalSupply())
-
-        const actionArgs = [
-          {
-            actionType: ActionType.Exercise,
-            owner: buyer,
-            sender: buyer,
-            asset: ethCall.address,
-            vaultId: '0',
-            amount: createScaledUint256(optionsAmount, 18),
-            index: '0',
-            data: ZERO_ADDR,
-          },
-        ]
-
-        await ethCall.approve(marginPool.address, createScaledUint256(optionsAmount, 18), {from: buyer})
-        await controller.operate(actionArgs, {from: buyer})
-
-        // keep track of balances after
-        const ownerWethBalanceAfter = new BigNumber(await weth.balanceOf(buyer))
-        const marginPoolWethBalanceAfter = new BigNumber(await weth.balanceOf(marginPool.address))
-        const ownerOtokenBalanceAfter = new BigNumber(await ethCall.balanceOf(buyer))
-        const marginPoolOtokenSupplyAfter = new BigNumber(await ethCall.totalSupply())
-
-        const payout = (strikePriceChange * optionsAmount) / expirySpotPrice
-
-        // check balances before and after changed as expected
-        assert.equal(
-          ownerWethBalanceBefore.plus(createScaledUint256(payout, (await weth.decimals()).toNumber())).toString(),
-          ownerWethBalanceAfter.toString(),
-        )
-        assert.equal(
-          marginPoolWethBalanceBefore.minus(createScaledUint256(payout, (await weth.decimals()).toNumber())).toString(),
-          marginPoolWethBalanceAfter.toString(),
-        )
-        assert.equal(
-          ownerOtokenBalanceBefore.minus(createScaledUint256(optionsAmount, 18)).toString(),
-          ownerOtokenBalanceAfter.toString(),
-        )
-        assert.equal(
-          marginPoolOtokenSupplyBefore.minus(createScaledUint256(optionsAmount, 18)).toString(),
-          marginPoolOtokenSupplyAfter.toString(),
-        )
-      })
     })
   })
 })
